+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# HybridMethod Library Change Log
+# Classicist Library Change Log
+
+## [1.0.1] - 2025-06-23
+### Added
+- Support for class properties via the `@classproperty` decorator.
+
+### Changed
+- Renamed the library to Classicist, available as `classicist` from PyPI.
 
 ## [1.0.0] - 2025-05-29
 ### Added